replace innerHTML with using parserUtils
This requires at least Firefox 14. In-item scripts are disabled, in-item
styles and external embeds are still enabled (may be disabled later).

Fixes #48 by disabling the in-item JS. However, #47 is likely not fixed
due to autoplay in <audio> and <video>. May lead to performance changes.
Affects #9 as relative links are now disabled completely.