do not check for a livemark by itemHasAnnotation
Using annotaion is a quick-and-dirty way to check whether an item is a
livemark without converting every function that needs this information
to async. On the other hand, it requires synchronous DB accesson the
main thread which may cause serious latency. Also this does not seem to
be an official API and thus may break at any moment. Overall, using the
official async API seems to be more reliable and performant.