Retarget links instead of intercepting clicks.
Intercepting clicks on in-content links and opening a new tab manually
has several drawbacks. These include obtaining incorrect address if
baseURI is specified (which leads to relative links not working) and
incorrect interactions with advanced tab-management addons that do not
recognize these as child tabs (not easily fixed manually for all the
half-dozen extensions).

The approach used here is to forcibly set for every link the 'target'
attribute to '_blank'. Due to performance considerations this is done
lazily in the 'onclick' handler before the link is activated.