do not special-case opening the entry with Enter
With focus tracking and links correctly working Enter does not need to
be separately tracked. Activating the auto-focused header link with
Enter will do the same and with less unintuitive moments. For example,
opening the entry in a new tab with Enter will automatically switch
focus to it iff it is the standard preference.

Remove the VK_RETURN hotkey and the support functions left unused.