Don't check in StorageInternal.tagEntry() if entry is already tagged.
This check is redundant because in Firefox tags are linked to URLs,
not bookmarks. While there can be many bookmarks for entry's URL,
there can't be more than one tag.

The check could be theoretically useful if the tag was removed while
Brief was disabled and then added again, but that's what
Query.verifyBookmarksAndTags() is for.