Fix and improve the option to exclude certain feeds from global views.
This option used to affect just the Unread folder, which doesn't exist
any more. Let's reword it and make it exclude the feed from both
All Items and Today views. Arguably, that's more in line with the
use case it has been intended to solve, than it used to be before.