FeedUpdateService should just hand the parsed feed as it is.
FeedFetcher.handleResult() does some preparations before handing
the data to FeedProcessor. The intention was for it to map the
necessary data to Brief's container, so that FeedProcessor didn't
have to deal with native feed interfaces. However, we now use
various properties that aren't represented in the containers and
we end up accessing wrappedFeed to bypass them anyway.