Convert our storage wrapper API to Promises.
The basic part is of course to return promises that are resolved
when statements are completed. But using promises calls for more
changes. handleCompletion() and handleError() have become redundant
given Promise's onFulfill and onReject methods, so they can be
removed. executeAsync() now takes a single function as callback
for handling individual rows.