"Organize" button should always select home folder in Library
When Organize button is clicked, the Bookmarks window should be opened
with Brief's home folder selected. This is done by passing the home folder
ID as an argument when opening the window. However, the built-in code that
handles window arguments only expands the target folder - it doesn't expand
its ancestors, so often the user cannot see the home folder upon clicking
Organize. We need to do this ourselves via an overlay.