Bug in removing day headings breaks loading entries in headlines view.
When entries are removed from a view, day headings are removed in
entry removal callbacks if the entry was the last one from that date.
However, when multiple entries are being removed, we must check if
the heading hasn't already been removed by another callback before
attempting to do it, because of the asynchronous nature of removal.