Tweak skin to allow for folder name overflow
If the folder name is too long, it used to cause horizontal scrolling
in the sidebar. As it is located in one container with a flex spacer,
use CSS flex instead of XUL flex to separate flex-grow and flex-shrink
into different elements. Also tweak the twisty's margins a bit to avoid
wasting space and tweak its padding so that open and closed folder have
equal amount of space for the title.