disable opening a release notes tab on update
Opening popups the user does not request is really not a nice behaviour,
even on the part of an addon. Also, it's far too easy for the tab to get
lost in the lots of tabs a user already has and just increase memory
consumption until he finds it.

On the other hand, most new versions are bugfixes and thus the user
probably does not need to know the exact list of bugs that do not exist
anymore. And if he really wants to see the release notes, they're
available on the addon page.

Finally, even if there are some new possibilities the user should be
aware of, it would probably be preferable to create the hints directly
in the Brief tab to ensure the user is in the correct context.

Fixes #173 (opening a tab for nonexistent release notes page on
brief.mozdev.org).