Avoid layout reflow when selecting a feed entry
Even though the modified border width and padding did cancel each other
a full layout recalculation was necessary. This is rather expensive when a
lot of entries are loaded.
This patch avoids the problem by adding a second div with a border that
is invisible by default. If an entry is selected, the border's color is updated.
See https://developer.mozilla.org/en-US/docs/Tools/Performance/Scenarios/Animating_CSS_properties#CSS_property_cost
for some details on the costs of a CSS property.