API.jsm: add StarUI
Note that the panel may be positioned badly if colliiding with screen
edges. The popup positioning used by Firefox cannot be used as-is
because rect-based positioning, as opposed to element-based, does not
support the arrow on the panel. This is a temporary little hack as all
this will need to go anyway with WebExtensions.