clone feed list into WebExtension storage
The feed list is duplicated into both `storage.local` and
`storage.sync`. The favicons and empty fields are both dropped to keep
the data size reasonable (especially for sync with 100kB limit). The
estimated data size is 500B/feed, so up to ~200 feeds can be stored in
the sync storage.