try to workaround the "lost home folder" problem
There's a well known Brief problem that in some cases (no way to
reliably reproduce known) Brief may accidentally lose sight of its home
folder, create a new one and require all feeds to be moved there by
hand, as if the original home folder was not found under its ID.

At least in one case it's known for sure that the home folder still
existed under the original ID. There are bugs in Mozilla bugzilla with
intermittent exceptions in getItemTitle (825739, 1081771) too.

In this commit `checkHomeFolder` is modified to retry calling
`getItemTitle` for up to 2 seconds (with 100ms intervals). This sounds
long, but note that the full 2-second interval is expected in *only one*
case - if Brief's home folder has been deleted - which is certainly
possible, but rare, and does not require *immediate* reaction.

This workaround is expected to be useful for Brief 2.4 series only, as
the WebExtension version won't be able to access Live Bookmarks anyway.