drop the deprecated Date.toLocaleFormat
This change may slightly affect date formatting, making it more natural
in some locales as Brief no longer hardcodes "day month year" as the
date order which may be not the one commonly used.

The locale is manually set to `navigator.language` as the default one is
often not correct (I was unable to make en-US Nightly on Linux use en-GB
with any available settings).