detect Brief downgrade and clear IDB for consistency
The migration code assumes and tries to ensure that all entries present
in the IndexedDB match the ones stored in `brief.sqlite`. Still, it may
become invalid in some cases.

First, some actions' modifications may be lost if Brief is disabled
right after a user performs some actions in it. However, this is a race
that involves the user and he's rather unlikely to win it.

Second, some actions' modifications may be lost on Firefox crash.
However, losing the latest changes on crash is an expected effect.

Third, some fetched entry updates may be lost in both of these cases.
This is considered acceptable as they will be fetched again in the pure
WE version.

Fourth, if a user downgrades Brief to a pre-hybrid version (lower than
2.4b1), the databases will become stale as the modifications are not
transferred to them.

This commit implements protection against the fourth scenario. Brief
detects a downgrade because all old versions used to set the
`lastVersion` pref to the Brief version if the current one is lower.
With this commit Brief sets it to a 'webextension-migration-done' marker
that's considered a lower version. If the marker is not detected, Brief
assumes that the entries may be arbitrarily stale and clears them to get
into a known state.