Hook "Subscribe Now" button in the default UI
This commit is a small pile of hacks designed to let users continue
using the "Subscribe Now" button on the feed preview page. It intercepts
the click when a Brief's item is selected, and sends a subscribe message
directly. It also creates an additional Brief item in the menu in case
there is no one in the user's prefs.