make the query engine more generic
The previous implementation was ad-hoc and required patching on every
new request type. Now the query-to-IDB mapping is generic and the ad-hoc
parts are isolated to `_guessIndexToUse` and `_possibleValues`, but even
incorrect values returned from these functions should make queries
slower (possibly MUCH slower) but still correctly perform all
discrete-values filtering.

Note that the sort/start/end/offset/limit filtering does not work in
some cases as it requires every cursor to be already sorted on the
required field (and thus requires the sort field to be indexed). There
are also no FTS hooks for now.