gracefully deal with duplicate URLs
If the items have no GUIDs and duplicate URLs (which is in general not
supported, but should behave in a somewhat reasonable way), try to match
them with recent items with the same URL from the DB in order. Would be
better to try matching content, but the current architecture does not
support that.