icons: add brief.svg (including #pulsing animation)
The SVG icon was roughly traced with Inkscape from the largest icon
`brief-icon-64.png`. No gradients/shadows/2.5D in the SVG, however, so
it's too rough for the places where it can be displayed at high detail
level, but should be ok for the pageAction icon (pulsing version for
feed preview pages).