pageAction: add animation on feed preview pages
The pageAction icon is an SVG that's animated (pulses fading in and out)
on feed preview pages to help the user find it and subscribe. It would
be better to integrate with the preview page, but WebExtensions don't
have the required permissions. Modal popups, forcibly redirecting the
user to some Brief page or subscribing right away would be too intrusive
in my opinion.