Migrator: integrate mergeFeeds and feedID mapping
Migration should be able to merge two feed lists if there are two in the
two storages, remove duplicates and preserve the folder structure. The
current implementation keeps the folder structure for the feed list
which was last updated and keeps the unique feeds and the folders
containing them for the other feed list (but drops the empty folders for
the mix-in list). It does not try to put feeds into the "guessed
correct" folders, however. The feed IDs are always modified for the
entries to be migrated (this is always done during migration).