package.json: add a list of eslint deps
Add a basic `package.json` file with dev-dependencies on eslint and all
plugins used. Note that Brief does not use it for build, npm is used for
dev-dependencies only. The lockfile is not committed for now.

This version installs the `babel-root-slash-import` import resolver as
it requires `node_modules` which npm will create anyway.