FeedView: avoid accessing FeedList for the feed list
The FeedList list of feeds may be wrong (if the FeedView is in preview
mode) and a layering violation (cyclic dependency). Cache the feed list
in the FeedView and update on its own (only in db-using mode). The
update is idempotent, so we don't care about extra updates in preview
mode.