request-monitor: add webRequest-based feed detection
The feeds served with the correct MIME types since Firefox 64 cause a
download prompt to appear (instead of a feed preview which had been
removed). A content script cannot intercept them, so this monitor checks
the MIME type and in the case of a feed redirects straight to the
Brief's preview mode.

Hopefully content handlers in WebExtensions (mozilla bug 1457500) will
make this part of code obsolete.

Fixes #411.