preview: do not use Brief feed preview before Fx65
The feed preview detection has a few false positives, for example during
a discarded tab restore, that are not critical for show-an-icon
workflow, but break in the redirect-to-preview one. The redirect from
feed preview also looks ugly due to a flicker of half-displayed Firefox
feed preview.

Taking into account that the only supported version with the Firefox
previews is Firefox ESR 60 and it's going out of support in a few
months, this patch limits the new Brief feed preview mode to Firefox
65+. This leaves Firefox up to 63 covered by the Firefox feed previews
and Firefox 64+ covered by the Brief feed previews. For some early
Firefox 64 nightlies both modes are active, which may result in a few
bugs that are not critical for bisection if needed.