Database.init: ignore persistent db if version<10
As found in #458 / mozilla bug 1556274, after trying to open the
persistent database and failing due to version 0 the database is in a
"ghost" state. It still does not exist, but opening the Storage
Inspector tab of developer tools causes it to appear with a version of 1
and no object stores breaking the Brief's migration source study.

As Brief has never used database versions below 10 (all versions are
multiples of 10), it can just ignore any database with version lower
than 10 as if it did not exist.

Fixes #458.