request-monitor: add content sniffer for feed tags
Since Firefox dropped its feed-related code in version 64 Brief has been
using a content script to inject into every page and check whether the
page is actually a feed. This kind of detection is basically required as
a significant number of servers fail to serve RSS/ATOM with the correct
MIME-type, and also pretty cheap as Brief has to inject its content
script anyway to detect links to feeds in the page.

This trick works for most pages. However, some websites use an
XSLT-based rendering for their feeds by specifying an `xml-stylesheet`
processing instruction that causes the browser to rewrite the XML. This
means that by the time content sctipt is loaded the original markup is
no longer there, so one cannot detect a feed at that stage.

To deal with this (and also speed up feed detection for other pages by
doing it earlier in the pipeline) this commit introduces a feed sniffer
modeled after the legacy Firefox sniffer. It parses the first 512 bytes
of the response and, if it looks like a valid XML and has a root tag
specific to feeds (and contains the RSS namespace URI in the RSS 1.0
case), triggers the redirect directly from the request monitor.