EntryView: cache date/time formatters
`toLocaleString` runs the formatter selection procedure on every call
(it caches only the "no options" formatter). Cache formatters manually.

Technically this causes a minor regression: changes in the browser
language require a page reload to activate (instead of affecting all
newly inserted items). If this ever becomes a problem, a
`languagechange` listener can be added to rebuild the formatters.

This change shaves off about 60% of the time spent inserting the items
on initial preview load.