Consistently use browser.i18n language
Make the language used for date formatting and document lang attribute
consistent with the language used for strings by using browser.i18n for
both mechanisms (instead of something relying on navigator.language
which is a different setting).

Also fixes the problems when navigator.language is not specified, see #487 and #274.